blip2/requirements.txt
transformers
pillow
requests
accelerate>=0.26.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0
--extra-index-url https://download.pytorch.org/whl/cu121